requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.3.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytils==0.3
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.2.0
